Fixed bug

+++ ExpenseAPI/src/main/java/com/example/expenseapi/repository/CurrencyRepository.java
@@ -3,4 +3,4 @@ package com.example.expenseapi.repository;
 import com.example.expenseapi.pojo.Currency;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CurrencyRepository extends CrudRepository<Currency, Integer> {}
+public interface CurrencyRepository extends CrudRepository<Currency, Long> {}
